refactored component with movie details
