+++ src/layouts/MoviesDetails/index.js
@@ -1,121 +1,34 @@
 import { useSelector } from "react-redux";
-import {
-  selectBackdropSizes,
-  selectPosterSizes,
-  selectProfileSizes,
-} from "../../features/configurationSlice";
-import {
-  selectMovieCredits,
-  selectMovieDetails,
-  selectMovieDetailsStatus,
-} from "../../features/movies/movieDetails/movieDetailsAndCreditsSlice";
-import { APIImageUrl } from "../../features/getAPI";
-import {
-  Background,
-  CastContent,
-  Content,
-  CrewContent,
-  List,
-  MainInfo,
-  Poster,
-  Raiting,
-  RaitingCap,
-  RaitingNumber,
-  RaitingWrapper,
-  StyledStar,
-  SubdivTitle,
-  TileContent,
-  Title,
-  Votes,
-  Wrapper,
-} from "./styled";
-import { Tile } from "../Tile/index";
+
+import BackgroundImageContent from "./BackgroundImageContent";
+import CreditsList from "./CreditsList";
 import Error from "../Error/index";
+import { Tile } from "../Tile/index";
+import { selectMovieDetailsStatus } from "../../features/movies/movieDetails/movieDetailsAndCreditsSlice";
+import { selectPosterSizes } from "../../features/configurationSlice";
+
+import { Content, TileContent } from "./styled";
 
 export const MoviesDetails = () => {
-  const movieDetails = useSelector(selectMovieDetails);
-  const movieCredits = useSelector(selectMovieCredits);
-  const backdropSizes = useSelector(selectBackdropSizes);
-  const posterSizes = useSelector(selectPosterSizes);
-  const profileSizes = useSelector(selectProfileSizes);
-  const movieDetailsStatus = useSelector(selectMovieDetailsStatus);
+   const posterSizes = useSelector(selectPosterSizes);
+   const movieDetailsStatus = useSelector(selectMovieDetailsStatus);
 
-  return (
-    <>
-      {movieDetailsStatus === "error" ? (
-        <Error />
-      ) : (
-        <>
-          {movieDetails && movieDetails.backdrop_path ? (
-            <Background>
-              <Wrapper>
-                <Poster
-                  source={`${APIImageUrl}/${backdropSizes ? backdropSizes[3] : ""}${
-                    movieDetails.backdrop_path
-                  }`}
-                  alt="movie poster"
-                />
-                <MainInfo>
-                  <Title>{movieDetails.title}</Title>
-                  <RaitingWrapper>
-                    <Raiting>
-                      <StyledStar />
-                      <RaitingNumber>
-                        {movieDetails.vote_average?.toFixed(1)}
-                        <RaitingCap>/10</RaitingCap>
-                      </RaitingNumber>
-                    </Raiting>
-                    <Votes>{movieDetails.vote_count} votes</Votes>
-                  </RaitingWrapper>
-                </MainInfo>
-              </Wrapper>
-            </Background>
-          ) : (
-            ""
-          )}
-          <Content>
-            <TileContent>
-              <Tile movieDetails posterSizes={posterSizes} />
-            </TileContent>
-            <CastContent>
-              <SubdivTitle>Cast</SubdivTitle>
-              <List>
-                {movieCredits &&
-                  movieCredits.cast &&
-                  movieCredits.cast.map(({ id, profile_path, name, character }) => (
-                    <Tile
-                      person
-                      profileSizes={profileSizes}
-                      key={id}
-                      id={id}
-                      profile_path={profile_path}
-                      name={name}
-                      character={character}
-                    />
-                  ))}
-              </List>
-            </CastContent>
-            <CrewContent>
-              <SubdivTitle>Crew</SubdivTitle>
-              <List>
-                {movieCredits &&
-                  movieCredits.crew &&
-                  movieCredits.crew.map(({ id, profile_path, name, job }) => (
-                    <Tile
-                      person
-                      profileSizes={profileSizes}
-                      key={id}
-                      id={id}
-                      profile_path={profile_path}
-                      name={name}
-                      job={job}
-                    />
-                  ))}
-              </List>
-            </CrewContent>
-          </Content>
-        </>
-      )}
-    </>
-  );
+   return (
+      <>
+         {movieDetailsStatus === "error" ? (
+            <Error />
+         ) : (
+            <>
+               <BackgroundImageContent />
+               <Content>
+                  <TileContent>
+                     <Tile movieDetails posterSizes={posterSizes} />
+                  </TileContent>
+                  <CreditsList cast title="Cast" />
+                  <CreditsList crew title="Crew" />
+               </Content>
+            </>
+         )}
+      </>
+   );
 };
